test: add post share test and delete toggle post type test
